Moved whole 'Add To Case' button into the separate component rather than just the dropdown. Disabled button after option selected
